Initial commit
Contains three vocal exercises which
can be loaded with Performous by
adding the path to this folder.

Also includes the mscz file which was
used to generate the mp3 file for each
vocal exercise. Not needed for
Performous, but may be useful
if the register of the exercise
is too high or low for you.